requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
sshtunnel==0.4.0
paramiko==3.0.0
PyMySQL==1.1.0
cryptography==41.0.7
email-validator==2.1.0
WTForms==3.1.1
stripe==7.8.0
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
docker==7.0.0
python-socketio==5.10.0
eventlet==0.33.3
